removing header for LHEPtFilter
